eidos/repo/pyproject.toml
[project]
name = "eidos-brain"
version = "0.4.7"
description = "Eidos Brain - Pattern matching engine for anomaly detection"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "streamlit>=1.25.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-pubsub>=2.17.0",
    "python-multipart",
    "requests",
    "websockets"
]

[project.scripts]
eidos = "eidos_brain.service.daemon:main"
eidos-api = "eidos_brain.service.api:main"
eidos-demo = "eidos_brain.demo.demo_runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eidos_brain"]

[tool.ruff]
line-length = 120

[tool.pyright]
include = ["src"]
